refactor: cleaned up directory structure and unused files
