+++ frontend/proxySetUp.js
@@ -1,3 +1,6 @@
+/**
+ * Used to solve CORS errors.
+ */
 const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function (app) {
