+++ frontend/src/components/shared/ToastSettings.jsx
