+++ frontend/src/components/unitHomePage/UnitCard.jsx
@@ -23,7 +23,7 @@ import {
     useToast,
 } from '@chakra-ui/react';
 import { BsTrash } from 'react-icons/bs';
-import getToastSettings from '../ToastSettings'
+import getToastSettings from '../shared/ToastSettings'
 import { IoEllipsisHorizontalSharp } from 'react-icons/io5';
 import { BrowserRouter, Routes, Route, useNavigate } from 'react-router-dom';
 import { MockAuth } from '../../mockAuth/mockAuth';
