+++ frontend/src/pages/UnitHomePage.jsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { useAuth0 } from '@auth0/auth0-react';
 import UnitCard from '../components/unitHomePage/UnitCard';
 import '../App.css';
-import getToastSettings from '../components/ToastSettings';
+import getToastSettings from '../components/shared/ToastSettings';
 import { MockAuth } from '../mockAuth/mockAuth';
 
 // Chakra imports
