--- frontend/src/sampleData/analytics.json
@@ -1,38 +1,0 @@
-[
-  {
-    "personality title": "Belbin team roles",
-    "description": "Personalities using Belbin team roles",
-    "categorical data": {
-      "title": "Belbin team roles distribution",
-      "x label": "Personality type",
-      "y label": "Number of students",
-      "x": ["thinking", "action", "people"],
-      "y": [50, 50, 25]
-    }
-  },
-  {
-    "personality title": "Student Effort",
-    "description": "How much effort is put to a unit and marks obtained",
-    "categorical data": [{
-      "title": "Average hour commitment to the unit",
-      "x label": "Average hour commitment",
-      "y label": "Number of students",
-      "x": ["0-4", "4-8", "8-12", "12+"],
-      "y":[5, 10, 20, 30, 15]
-    },
-    {
-      "title": "Average assignment mark distribution",
-      "x label": ["Average assignment marks"],
-      "y label": ["Number of students"],
-      "x": ["N", "P", "C", "D", "HD"],
-      "y":[10, 10, 20, 30, 15]
-    }],
-    "numerical data": {
-      "title": "Distribution of marks per hour",
-      "x label": "Marks per hour",
-      "y label": "Number of students",
-      "x": [0.2, 0.5, 0.6, 0.8, 1.2],
-      "y": [10, 15, 17, 22, 7]
-    }
-  }
-]
