--- frontend/src/sampleData/unit.json
@@ -1,6 +1,0 @@
-{
-  "unitId": "ENG1003",
-  "unitFaculty": "Engineering",
-  "labs": ["001", "002", "003"],
-  "teachers": ["EvanSmith@teacher.monash.edu", "JaneDoe@teacher.monash.edu"]
-}
